pyproject.toml
[project]
name = "mcp-proxy-for-llama-cpp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp==3.12.4",
    "dacite==1.9.2",
    "mcp==1.9.2",
    "pyyaml==6.0.2",
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]

[dependency-groups]
dev = [
    "ruff==0.11.12",
]
